+++ pochak/src/main/java/com/apps/pochak/user/service/UserService.java
@@ -3,9 +3,9 @@ package com.apps.pochak.user.service;
 import com.apps.pochak.common.BaseException;
 import com.apps.pochak.user.domain.User;
 import com.apps.pochak.user.dto.UserFollowersResDto;
+import com.apps.pochak.user.dto.UserFollowingsResDto;
 import com.apps.pochak.user.dto.UserUpdateRequestDto;
 import com.apps.pochak.user.dto.UserUpdateResDto;
-import com.apps.pochak.user.dto.UserFollowingsResDto;
 import com.apps.pochak.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,8 +15,6 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static com.apps.pochak.common.BaseResponseStatus.*;
-import static com.apps.pochak.common.BaseResponseStatus.DATABASE_ERROR;
-import static com.apps.pochak.common.BaseResponseStatus.NULL_USER_HANDLE;
 
 
 @Service
@@ -79,7 +77,7 @@ public class UserService {
     @Transactional
     public UserUpdateResDto updateUserProfile(String updatedUserHandle, UserUpdateRequestDto requestDto) throws BaseException {
         try {
-            User userWithUserHandle = userRepository.findUserWithUserHandle(updatedUserHandle);
+            User userWithUserHandle = userRepository.findUserByUserHandle(updatedUserHandle);
 
             // profileImage와 한줄 소개는 null 값 가능하게 설정.
             if (requestDto.getName().isBlank()) {
@@ -110,7 +108,7 @@ public class UserService {
 
     public Boolean checkHandleDuplicate(String handle) throws BaseException {
         try {
-            userRepository.findUserWithUserHandle(handle);
+            userRepository.findUserByUserHandle(handle);
             return true; // 중복됨
         } catch (BaseException e) {
             if (e.getStatus().equals(INVALID_USER_ID)) {
